tidying and code organization
