+++ src/App.jsx
@@ -14,7 +14,7 @@ import './App.scss'
 import logoWhite from './assets/logo/TollGate_Logo-C-white.png';
 
 export const App = () => {
-  const { t: i18n, i18n: i18n_ } = useTranslation();
+  const { t } = useTranslation();
   const [method, setMethod] = useState('cashu');
   const [tollgateDetails, setTollgateDetails] = useState(null);
   const [loading, setLoading] = useState(true);
@@ -25,7 +25,7 @@ export const App = () => {
 
   // Initial data fetch on mount
   useEffect(() => {
-    const initialFetch = async () => {
+    const fetch = async () => {
       setLoading(true);
       const response = await fetchTollgateData();
       
@@ -39,7 +39,7 @@ export const App = () => {
       setLoading(false);
     };
     
-    initialFetch();
+    fetch();
     
     // Cleanup on unmount
     return () => {
@@ -83,56 +83,77 @@ export const App = () => {
     };
   }, [error, tollgateDetails]);
 
-  return <div id="tollgate-captive-portal" className="tollgate-captive-portal">
-    <Background />
-    <div className="tollgate-captive-portal-interface">
-      <div className="tollgate-captive-portal-header">
-        <img src={logoWhite} alt="TollGate Logo"></img>
-      </div>
-      <div className="tollgate-captive-portal-content">
-        <div className="tollgate-captive-portal-content-container">
-          <div className="tollgate-captive-portal-tabs" aria-label="Tab Navigation">
-            <button 
-              onClick={() => setMethod('cashu')}
-              data-active={method === 'cashu'}
-              className="tollgate-captive-portal-tabs-tab tollgate-captive-portal-tabs-tab-cashu ellipsis" 
-              label={i18n('cashu_tab')} 
-              id="tab-cashu" 
-              aria-controls="tab-cashu">
-              {i18n('cashu_tab')}
-            </button>
-            <button 
-              onClick={() => setMethod('lightning')}
-              data-active={method === 'lightning'}
-              className="tollgate-captive-portal-tabs-tab tollgate-captive-portal-tabs-tab-lightning ellipsis" 
-              label={i18n('lightning_tab')} 
-              id="tab-lightning" 
-              aria-controls="tab-lightning">
-              {i18n('lightning_tab')}
-            </button>
+  return (
+    <div id="tollgate-captive-portal" className="tollgate-captive-portal">
+      <Background />
+
+      <div className="tollgate-captive-portal-interface">
+        <Header />
+
+        <div className="tollgate-captive-portal-content">
+          <div className="tollgate-captive-portal-content-container">
+
+            <div className="tollgate-captive-portal-tabs" aria-label="Tab Navigation">
+              <Tab type="cashu" method={method} setMethod={setMethod} />
+              <Tab type="lightning" method={method} setMethod={setMethod} />
+            </div>
+
+            <div className="tollgate-captive-portal-view">
+              {loading && <Loading />}
+
+              {!loading && error && <div className="tollgate-captive-portal-error">
+                <Error label={error.label} code={error.code} message={error.message} />
+              </div>}
+
+              {!loading && !error && method === 'cashu' && <Cashu tollgateDetails={tollgateDetails.value} />}
+              {!loading && !error && method === 'lightning' && <Lightning tollgateDetails={tollgateDetails.value} />}
+            </div>
+
           </div>
-
-          <div className="tollgate-captive-portal-view">
-            {loading && <div className="tollgate-captive-portal-loading">
-              <span className="spinner big"></span>
-              Loading
-            </div>}
-
-            {!loading && error && <div className="tollgate-captive-portal-error">
-              <Error label={error.label} code={error.code} message={error.message} />
-            </div>}
-
-            {!loading && !error && method === 'cashu' && <Cashu tollgateDetails={tollgateDetails.value} />}
-            {!loading && !error && method === 'lightning' && <Lightning tollgateDetails={tollgateDetails.value} />}
-          </div>
-
         </div>
+
+        <Footer />
       </div>
-      <div className="tollgate-captive-portal-footer">
-        <p><Trans i18nKey="powered_by" components={{ 1: <a href="https://tollgate.me/" target="_blank" rel="noreferrer"></a> }} /></p>
-      </div>
+
     </div>
-  </div>;
+  );
+}
+
+const Header = () => {
+  const { t } = useTranslation();
+
+  return <div className="tollgate-captive-portal-header">
+    <img src={logoWhite} alt="TollGate Logo"></img>
+  </div>
+}
+
+const Tab = ({ type, method, setMethod }) => {
+  const { t } = useTranslation();
+
+  return <button 
+    onClick={() => setMethod(type)}
+    data-active={method === type}
+    className={`tollgate-captive-portal-tabs-tab tollgate-captive-portal-tabs-tab-${type} ellipsis`} 
+    label={t(`${type}_tab`)} 
+    id={`tab-${type}`}
+    aria-controls={`tab-${type}`}>
+    {t(`${type}_tab`)}
+  </button>
+}
+
+const Loading = () => {
+  const { t } = useTranslation();
+
+  return <div className="tollgate-captive-portal-loading">
+    <span className="spinner big"></span>
+    {t('loading')}
+  </div>
+}
+
+const Footer = () => {
+  return <div className="tollgate-captive-portal-footer">
+    <p><Trans i18nKey="powered_by" components={{ 1: <a href="https://tollgate.me/" target="_blank" rel="noreferrer"></a> }} /></p>
+  </div>
 }
 
 export default App
