+++ src/helpers/tollgate.js
@@ -1,6 +1,3 @@
-import { nip19, getPublicKey, getEventHash, finishEvent, signEvent } from 'nostr-tools';
-import { getDecodedToken } from '@cashu/cashu-ts';
-
 // Get the base URL dynamically
 export const getTollgateBaseUrl = () => {
   // Get the current host (without port)
@@ -61,42 +58,6 @@ export const fetchTollgateData = async () => {
       message: 'TollGate could not connect to its relay. Contact the network administrator or try again later.'
     };
   }
-};
-
-// Safely extract proofs from a decoded token
-export const extractProofsFromToken = (decodedToken) => {
-  const proofs = [];
-  
-  try {
-    // Handle token array format
-    if (decodedToken.token && Array.isArray(decodedToken.token)) {
-      decodedToken.token.forEach(t => {
-        if (t.proofs && Array.isArray(t.proofs)) {
-          proofs.push(...t.proofs);
-        }
-      });
-    }
-    // Handle single token format
-    else if (decodedToken.token && decodedToken.token.proofs) {
-      proofs.push(...decodedToken.token.proofs);
-    }
-    // Handle direct token structure
-    else if (decodedToken.proofs && Array.isArray(decodedToken.proofs)) {
-      proofs.push(...decodedToken.proofs);
-    }
-    // Handle token with tokens array
-    else if (decodedToken.tokens && Array.isArray(decodedToken.tokens)) {
-      decodedToken.tokens.forEach(t => {
-        if (t.proofs && Array.isArray(t.proofs)) {
-          proofs.push(...t.proofs);
-        }
-      });
-    }
-  } catch (error) {
-    console.error('Error extracting proofs:', error);
-  }
-  
-  return proofs;
 };
 
 // Helper function to format data size in user-friendly units
@@ -243,199 +204,4 @@ export const calculatePurchasedAllocation = (sats, metric, price, stepSize, i18n
   }
 
   return purchasedStr;
-};
-
-// When token is pasted, extract proofs and calculate total value
-export const validateToken = (token = '') => {
-  try {
-    if (!token.trim()) {
-      return {
-        status: 0,
-        code: 1,
-        label: 'Token missing',
-        message: 'No token provided.'
-      };
-    }
-
-    // Basic validation - Cashu tokens should start with "cashu"
-    if (!token.trim().startsWith('cashu')) {
-      return {
-        status: 0,
-        code: 2,
-        label: 'Invalid token format',
-        message: 'Cashu tokens should start with "cashu".'
-      };
-    }
-
-    // Attempt to decode the token using the getDecodedToken method
-    let decodedToken = null;
-    try {
-      decodedToken = getDecodedToken(token.trim());
-      if (!decodedToken) {
-        return {
-          status: 0,
-          code: 3,
-          label: 'Invalid Cashu Token',
-          message: 'Token could not be decoded.'
-        };
-      }
-    } catch(err) {
-      return {
-        status: 0,
-        code: 3,
-        label: 'Invalid Cashu Token',
-        message: 'Token could not be decoded.'
-      };
-    }
-
-    // Extract proofs from the token
-    const proofs = extractProofsFromToken(decodedToken);
-    if (!proofs || proofs.length === 0) {
-      // If we couldn't extract proofs, still accept the token but show no value
-      return {
-        status: 1,
-        value: {
-          isValid: true,
-          hasProofs: false
-        }
-      };
-    }
-
-    // Calculate the sum of proof values
-    const totalAmount = proofs.reduce((sum, proof) => {
-      const proofAmount = Number(proof.amount || 0);
-      return sum + proofAmount;
-    }, 0);
-
-    // Return the token value with proof details
-    return {
-      status: 1,
-      value: {
-        isValid: true,
-        hasProofs: true,
-        amount: totalAmount,
-        proofCount: proofs.length,
-        unit: 'sats'
-      }
-    };
-  } catch (error) {
-    console.error('Error decoding token:', error);
-    return {
-      status: 0,
-      code: 4,
-      label: 'Token validation error',
-      message: error.message || 'Invalid token format.'
-    };
-  }
-};
-
-// Handle manual page close
-export const handleClosePage = () => {
-  // Set that the close button was clicked
-  setCloseButtonClicked(true);
-  
-  // Try closing the window
-  window.close();
-  
-  // If we're still here after a short delay, the close failed
-  setTimeout(() => {
-    // No need to show alert anymore since we'll show the message in UI
-    // The fact that this code runs means the window wasn't closed
-  }, 300);
-};
-
-// Update the handleSendToken function
-export const handleSendToken = async () => {
-  try {
-    setStatus('Sending token...');
-    
-    if (!deviceInfo || !tollgateDetails) {
-      throw new Error('Missing device information or TollGate details');
-    }
-    
-    // Get TollGate pubkey from event
-    const tollgatePubkey = tollgateDetails.pubkey;
-    
-    // Generate a random private key for signing
-    const privateKeyBytes = window.crypto.getRandomValues(new Uint8Array(32));
-    const privateKeyHex = Array.from(privateKeyBytes)
-      .map(b => b.toString(16).padStart(2, '0'))
-      .join('');
-    
-    // Generate the pubkey from private key using nostr-tools
-    let pubkey;
-    try {
-      pubkey = getPublicKey(privateKeyHex);
-    } catch (e) {
-      console.error('Error getting public key:', e);
-      throw new Error('Failed to generate keys for signing');
-    }
-    
-    // Create the Nostr event according to TIP-01 spec
-    const unsignedEvent = {
-      kind: 21000,
-      pubkey: pubkey,
-      content: "",
-      created_at: nostrNow(),
-      tags: [
-        ["p", tollgatePubkey],
-        ["device-identifier", deviceInfo.type, deviceInfo.value],
-        ["payment", token],
-      ],
-    };
-    
-    // Calculate the event hash (id)
-    const id = getEventHash(unsignedEvent);
-    
-    // Sign the event using signEvent, which is still available
-    const sig = signEvent(unsignedEvent, privateKeyHex);
-    
-    // Create a clean event object for sending
-    const event = {
-      ...unsignedEvent,
-      id,
-      sig
-    };
-    
-    console.log('Sending signed event:', event);
-    
-    // Send the event to the TollGate server
-    const baseUrl = getTollgateBaseUrl();
-    const response = await fetch(`${baseUrl}/`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(event),
-    });
-    
-    if (!response.ok) {
-      if (response.status === 402) {
-        throw new Error('Payment required. Your token was not accepted.');
-      } else {
-        throw new Error(`Server error: ${response.status}`);
-      }
-    }
-    
-    // Success!
-    setStatus('');
-    setShowSuccess(true);
-    
-    // Try to auto-close the page after 3 seconds
-    setTimeout(() => {
-      window.close();
-      // If we're still here after a short delay, show the close button
-      setTimeout(() => {
-        setShowCloseButton(true);
-      }, 500);
-    }, 3000);
-    
-    // Reset form
-    setToken('');
-    setTokenValue(null);
-    
-  } catch (error) {
-    console.error('Error sending token:', error);
-    setStatus(`Error: ${error.message}`);
-  }
 };
